+++ CHANGELOG.md
@@ -1,9 +1,16 @@
 # Changelog
 All notable changes to the Audio Player project will be documented in this file.
 
-## 3.4.2 - 2025-07-12
+## 3.5.0 - in progress
+### Added
+
+### Changed
+- refactored MusicController into service and mapper
+- migrated MusicMapper to use QueryBuilder
+
 ### Fixed
 - PHP 8.4 compatibility for nullable parameters
+
 
 ## 3.4.1 - 2023-12-11
 ### Fixed